+++ src/convertdate/__init__.py
@@ -38,7 +38,7 @@ from . import positivist
 from . import ordinal
 from . import utils
 
-__version__ = '2.3.0'
+__version__ = '2.3.1'
 
 __all__ = [
     'holidays',